Restructure assessment files * Move components to their own folder * Define pages components * Rename Component.intl.js to translations.intl.js
