+++ client/app/bundles/course/assessment/assessments-index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import ProviderWrapper from 'lib/components/ProviderWrapper';
 import storeCreator from './store';
-import PopupDialog from './containers/PopupDialog';
+import AssessmentIndexPage from './pages/AssessmentIndex';
 
 $(document).ready(() => {
   const mountNode = $('.new-btn')[0];
@@ -13,7 +13,7 @@ $(document).ready(() => {
     const store = storeCreator({ assessments: {} });
     const Page = () => (
       <ProviderWrapper store={store}>
-        <PopupDialog
+        <AssessmentIndexPage
           courseId={attributes.course_id}
           categoryId={attributes.category_id}
           tabId={attributes.tab_id}
